Fix up the code for torch on after recording has started
